Khepri: Tree-like replicated on-disk database library
Data are stored in a **tree structure**. Each node in the tree is referenced by
its path from the root node. A path is a list of Erlang atoms and/or binaries.
For ease of use, Unix-like path strings are accepted as well.

For **consistency and replication** and to manage data on disk, Khepri relies
on the [Ra library](https://github.com/rabbitmq/ra). The Ra library implements
the [Raft consensus algorithm](https://raft.github.io/). In Ra parlance,
Khepri is a state machine in a Ra cluster.

The initial prototype was written by Karl Nilsson <nkarl@vmware.com>.