Transactions: Supports transactions executed from an Erlang shell
What is special with an Erlang shell is that the code is evaluated with
`erl_eval`. If the code uses anonymous functions, they are sort of
loaded "inside" the `erl_eval` module. In other words they appear as
local functions of this module.

In the context of Khepri transactions, where we need to extract the
assembly code, `khepri_fun` would fail to extract anything because the
`erl_eval` module doesn't really contain the assembly code of whatever
it evaluates. Instead, the abstract form of an anonymous function is
stored in its environment (the `env` key in the proplist returned by
erlang:fun_info/1).

This patch takes that abstract form, compiles it and extract the
assembly from the compiler beam. The rest of the processing is
unchanged.

While here, we convert the `fun_info` proplist to a map and store that
in our #state{} record. It's much easier to use and allows pattern
matching on the function type in the newly added feature.