Transactions: Detect transaction nature when ReadWrite = auto
In other words, if the analysis of the code show that it calls
khepri_tx:put/{2,3} or khepri_tx:delete/1, the transaction is considered
a read/write one. It is the same as using `khepri:transaction(Fun, true)`.

Otherwise, the transaction is considered read-only. It is the same as
using `khepri:transaction(Fun, false)`.

This should make it easier to use transactions without having to specify
if it is read-only or read/write. Of course, it means that in `auto`
mode, read-only anonymous functions are analyzed: if the transaction is
known to be read-only, mentionning it will make it a bit faster by
skipping the analysis.

Note that the default value for ReadWrite is now `auto` instead of
`true`.