Test: Add property-based testing for khepri_path and khepri_machine
This testing is not that extensive at this time. For instance, the
Khepri state machine is tested with simple paths without any conditions.
However, this allowed us to discover a couple bugs in the way paths are
converted back and forth between their native and their string-based
forms. The bug fix will comme in a follow-up commit.