GitHub Actions: Cache Dialyzer PLT files
They are expensive to generate and they are perfect candidates for
caching because they do not change for a given version of Erlang.

The `rebar.config.script` file was reorganized because an earlier
version of that patch depended on a change in that file. This is no
longer needed, but I kept the reorganization to make it easier to extend
`rebar.config.script` in the future.