Rebar: Run Dialyzer on testsuites
In order to run Dialyzer on testsuites, we have to add `test` to the source
directories, even though this is semantically incorrect. We also need to
declare PropEr as a regular dependency, not a test-specific one.

The downsides are that PropEr becomes a regular dependency (even when
this library is used as a dependency and never tested) and test modules
are compiled by default...

Once and if Rebar grows support for Dialyzer on tests, we can get rid of all
this.

V2: The GitHub Actions test workflow is updated to take into account any
    modifications to `rebar.config*` in the Dialyzer PLT files cache
    key. Without that, it will miss any extra applications we declare
    in Dialyzer options.

V3: Run `rebar3 clean` instead of `rebar3 clean -a` before running
    Dialyzer. The reason is that `proper`'s clean script is missing when
    fetched from Hex.pm, breaking `rebar3 clean -a`.