khepri_utils: Handle unordered keys in `flat_struct_to_tree/1`
Map keys are unordered and the function didn't handle that. So if in
`maps:fold/3`, `[foo, bar]` was handled before `[foo]`, the code would
hit an assertion and crash.

Now it correctly merges `[foo]` node props with its existing child
nodes.

This was reported by @mkuratczyk while testing the integration of Khepri
in RabbitMQ.