khepri_path: Call realpath/1 in pattern_includes_root_node/1
This cleans any use of `?ROOT_NODE` or `?THIS_NODE` in the path.
Otherwise the function would return a different result if these macros
were prepended to `?STAR` or `?STAR_STAR` for instance, even though the
meaning of the path is exactly the same.

I'm still not very happy with this heuristic to determine if the root
node properties should be included in a get. Another solution would be
to require an explicit `?ROOT_NODE` in the path to include its
properties in the result.