khepri_fun: Fix extraction of Elixir anonymous functions
... written in the Elixir shell (`erl_eval` anonymous functions).

In the `erl_eval` anonymous function's env, the first element in the
tuple represents the function bindings. Erlang is using a list, but
Elixir is using a map.

`erl_eval_fun_to_asm()`, which does pattern matching to ensure there is
no bindings, expected an empty list. Now it accepts an empty map as well
to support the Elixir case.

Note that bindings are still unsupported.

Reported by @mkuratczyk.
Paired with @mkuratczyk.