Rebar: Remove hack to run Dialyzer on testsuites
In order to run Dialyzer on testsuites, I changed `extra_src_dirs` to
include the test directory as well. Unfortunately, after adding the
property-based testcases, I had to add PropEr to the dependencies...

The other solution is to explicitly run Dialyzer under the `test` Rebar
profile. I don't know any way to tell Rebar to use that profile by
default with Dialyzer, so one has to remember to use it on the command
line.

The `check` alias does not set any specific profile, so Dialyzer is
executed on the library only or the user must specify the `test` profile
on the command line. The problem with the latter is that it will affect
other commands such as `xref`.