GitHub Actions: Explicitly set `test` profile for Dialyzer
By default, Dialyzer runs only on the library source code. I want to run
it on the testsuites as well. I previously had to put in place a hack in
`rebar.config` to enforce that.

The huge drawback with that hack is that PropEr became a dependency,
even though it was not used at runtime.

Therefore, the hack is being removed. To keep running Dialyzer on the
testsuites, we must specify the `test` profile explicitly on the Rebar
command line.

The path to the Dialyzer PLT to cache is updated at the same time
because now, they will be under the `_build/test` directory.