Transactions: Allow `re` module
However, `re:version/0` is denied: it could return a different result on
different nodes.

For `re:compile/{1,2}`, it is more complicated. The compiled regex could
be very specific to a version of the PCRE engine and not portable to
another node. It is fine to compile a regex and use it in the same
function (transaction). But storing it in Khepri and using it on another
node than the one which compiled it could break. This API is still
allowed because there is no way for us to prevent a compiled regex to
come from outside a transaction function or to be stored by
`khepri_machine:put/3` for instance.