Rebar: Follow-up commit to the addition of rebar3_edoc_extensions
First, a quick explanation of the reason which was missing in the
initial commit: this Rebar 3 plugin is made from the `khepri_edoc_*`
modules. The plugin takes care of downloading the GitHub Markdown and
PrismJS files. That's why we can remove the EDoc modules and JS+CSS
files: everything was moved from this repository to the plugin.

This follow-up commit removes the GitHub Markdown and PrismJS files
which were not removed in the initial commit.