preemptively infer var_info annotations in pass1_process_instructions/3
In many cases we can infer the necessary typing of a variable based on
an instruction in a function body. For example, when we see

    {get_tuple_element, Src, Element, _Dst}

We know that the register Src must contain a tuple with at least
`Element + 1' elements (if it weren't, the code could not have been
initially compiled), so we can add

    {'%', {var_info, Src, [{type, {t_tuple, Element + 1, false, #{}}}]}}

To satisfy the beam_validator's type-checks. This commit removes the clause
of `handle_validation_error/3' which would correct these failures when
`compile:forms/2' would emit them: that clause no longer matches.