add test case for function with bitstring typing
parse_float/1 starts with a bs_start_match4 instruction that needs
to both accept match contexts and must be typed as a t_bitstring in
order to satisfy the beam_validator. t_bitstring cannot be inferred
in pass1_process_instructions/3 (see the parent commit), so it needs
to be handled after it is noticed in the error of compile:forms/2.
See the child commit for the necessary changes to compile
parse_float/1 in a khepri_fun.