add comments at exact locations when fixing validation errors
The implementation of add_comment_to_code/5 prior to this commit
placed comments into function code immediately following the function's
entry-point label, which is correct in most scenarios. Some validation
failures may need comments at other points in the function code, though.
Luckily, beam_validator points to the location of the failing instruction
in the failure triple. So this commit allows one to pass the location where
the comment should go end up in the function code.

Included with this new structure are two validation-failure handling clauses:
one may add type information to a bs_start_match4 instruction when the
validator is not satisfied by the current lack of typing. (I don't believe
you can infer the typing of a bs_start_match4 easily, we'd have to look
into the validator to see how it's detecting the 'needed' type.) The other
is an interesting case where an intermediary function needs an
'accepts_match_context' annotation. In this case, function A calls function
B with some variable that accepts a match context. Function B then calls
function C, which accepts a variable that must be a match context. Functions
A and C may have their variables properly annotated, but function B does not,
because it does not contain any bs_start_match* instructions. It may look
something like:

    {function, 'B', 1, 30,
      [
        {label, 29},
        {func_info, {atom, my_module}, {atom, 'B'}, 1},
        {label, 30},
        {call_only, 1, {f, 32}} %% function 'C'
      ]}

So in this case we ignore the FailingFun and instead look at the entry-label
in the failing call/call_only label and use that to find the function. Then
we annotate the {x,0} variable in that function to accept a match context,
which satisfies the validator.

This is currently hard-coded for a 1-arity function. More testing will be
needed to find an example of this where we can't blindly assume {x,0}.