khepri_fun: Add `select_tuple_arity` type annotation
The `select_tuple_arity` is used in the following code for instance:

    case Tuple of
        {ok, _, _}   -> true;
	{error, _}   -> false;
	{timeout, _} -> false
    end

In other words, the VM can simply verify the arity of the tuple to
decide between `true` and `false` if the atom in the first position was
implicitly verified earlier in the code.

`beam_validator` expects the annotation to specify a tuple without any
specific arity.