khepri_fun: Complete "Line" beam chunk decoding
The last part of the `read_tagged()` function was left unimplemented
until we find an actual use case. We got one with the `rabbit_amqqueue`
module in RabbitMQ which is long enough.

The "Line" beam chunk from this module is added as a testcase and the
rest of `read_tagged()` is implemented in this patch.

The `#lines{}` is not in a header for now, that's why the testcase
hard-codes the matching.

Fixes #53.