add test case functions with new arities for hard-coded register case
These test case functions are almost the same as the original
parse_float and parse_digits functions, except that they introduce
extra intermediary calls of parse_digits2 with additional bindings
prepended to the parameter list. These new bindings throw off the
arity of the bitstring match parameter `Bin` for which the original
validation fix tested by `allowed_bs_match_digit_parser_test` was
hard-coded.

These test cases also reveal some interesting behavior when we play
with the current code for fixing this validation error. This is the
original code which hard-codes a fix for a 1-arity function missing
the annotation on register {x,0} (the only argument):

    handle_validation_error(
      Asm,
      {_FailingFun,
       {{Call, 1 = _Arity, {f, EntryLabel}},
        _,
        no_bs_start_match2}},
      Error) when Call =:= call orelse Call =:= call_only ->
        Register = {x, 0},
        Comment = {'%', {var_info, Register, [accepts_match_context]}},
        Location = {'after', {label, EntryLabel}},
        add_comment_and_retry(Asm, Error, EntryLabel, Location, Comment);

Let's add a clause for a 2-arity function and hard-code the register
to {x,1}:

    handle_validation_error(
      Asm,
      {_FailingFun,
       {{Call, 2 = _Arity, {f, EntryLabel}},
        _,
        no_bs_start_match2}},
      Error) when Call =:= call orelse Call =:= call_only ->
        Register = {x, 1},
        Comment = {'%', {var_info, Register, [accepts_match_context]}},
        Location = {'after', {label, EntryLabel}},
        add_comment_and_retry(Asm, Error, EntryLabel, Location, Comment);

Initially, I expected this case to cover the new `parse_digits2/2` but
not `parse_digits2/3`. But In fact, `parse_digits2/3` needs no
annotations. This means that in order to fix this case, we don't need
to reach into the call-graph of `parse_digits2` and determine the
typings of the functions it calls. Instead we need to propagate the
`accepts_match_context` typing from `parse_float2/1` (`parse_digit2`'s
only caller).

This is good news because the validation error data does not include
any information about the eventual call to `parse_digits2/4`. We would
need to reach down the entire call-graph from `parse_float2/1` to
determine the types. Instead, the validation error contains enough
context to add the necessary annotations without ambiguity (see the
next commit).