rename lookup_function1 to find_function
This function is useful for finding a function by name and arity
(if defined) in a list of functions. We get a list of functions
either from the asm of the module being generated or from the
disasm from an existing module being disassembled, so this function
can be used by both `handle_validation_error/3' and `lookup_function/4`.