Documentation: Clarify the difference for triggers between Khepri and RDBMS
A new section to the documentation was added to clarify that Khepri
triggers are not the same as RDBMS ones, despite the same name. Triggers
in Khepri can't mess with the database content directly.

This fear comes from some feedback after the announcement of triggers.
The person who expressed that fear was reassured after knowing that
difference.

While here, fix a typo in the paragraph above the added section.

Fixes #58.