khepri_fun: improve coverage for infix operators
The compiler is smart enough to optimize away pretty much all the
instructions that were in `allowed_erlang_expressions_test/0'. In
fact, the asm for the begin/end block in that test came out to:

    {function, '-allowed_erlang_expressions_test/0-fun-0-', 0, 722}.
      {label,721}.
        {line,[{location,"test/tx_funs.erl",65}]}.
        {func_info,{atom,tx_funs},
                   {atom,'-allowed_erlang_expressions_test/0-fun-0-'},
                   0}.
      {label,722}.
        {test,has_map_fields,{f,723},{literal,#{a => b}},{list,[{atom,a}]}}.
        {move,{literal,#{a => c}},{x,0}}.
        return.
      {label,723}.
        {line,[{location,"test/tx_funs.erl",106}]}.
        {badmatch,{literal,#{a => b}}}.

This is a refactor of that test case which treats each case differently
so that the compiler doesn't think any of the calculations can be
discarded.

The case for division has been split out to another commit as it starts
involving instructions specifically used for floating point arithmetic.
See the child commit for more details.