khepri_fun: Cache compiled standalone function
The cache key is the anonymous function's module, function name and
arity, the options passed to `to_standalone_fun()`, plus the checksum of
the module holding it.

The cache value is the standalone function, the checksums of all modules
called directly or indirectly by the function including the modules
referenced by callbacks in the options map.

This cache allows to skip both passes and the compilation. A quick
benchmark showed the following rates for the same transaction executed
in a loop by 140 concurrent processes:
* without cache: 20k transactions per second
* with cache: 60k transactions per second

The transaction only did a `khepri_tx:put/2` call. As a comparison, a
direct call to `khepri_machine:put/3` leads to 80k insertions per
second.

V2: The second implementation uses the list of modules involved and
    their checksum to determine a cache hit. The initial implementation
    used a hash of the collected assembly between pass1 and pass2. This
    meant that we only skipped pass2 and the compilation. The
    performance reached with this first implementation was 40k
    transactions per second.

V3: Move the `to_standalone_fun()` options to the cache key. Options
    could lead to a different standalone function, so we must have
    separate cache entries for different options. Options were verified
    in the initial implementation, but different options would overwrite
    the existing cache entry instead of having a dedicated one.