Accept string-based paths in khepri_machine and khepri_tx
This was the case for the "simple API" in `khepri` only so far. But this
is convenient to be able to use string-based paths anywhere. It also
improves the consistency of the API.

As a consequence, `khepri_path:from_string/1` accepts both string-based
and regular paths and returns a regular path. Now unnecessary,
`khepri_path:maybe_from_string/1` and
`khepri_path:component_from_string/1` were removed.

`khepri_path:component_to_string/1` is still there but made internal.
It's not public anymore.