khepri_path: Change syntax of string-based paths
First, atoms are represented as `:atom`. Binaries are simpler and
represented as `binary` (the prefix and suffix are gone).

Then, atoms and binaries are percent-encoded in order to make them
compatible with URIs. For instance, if an atom or binary contains a
forward slash, it can be encoded as `%2F`.

Note that the empty atom `''` will be represented as `":"`, however the
empty binary `<<>>` can't be represented.

Here a some examples:

| String-based path | Regular path                             |
|-------------------|------------------------------------------|
| `"/:foo/:bar"`    | `[foo, bar]`                             |
| `":foo/:bar"`     | `[?THIS_NODE, foo, bar]`                 |
| `"./:foo/:bar"`   | `[?THIS_NODE, foo, bar]`                 |
| `"/foo"`          | `[<<"foo">>]`                            |
| `"/:AC%2FDC"`     | `['AC/DC']`                              |
| `"/glob*"`        | `[#if_name_matches{regex = "^glob.*$"}]` |
| `"/:foo/**"`      | `[foo, ?STAR_STAR]`                      |

The new syntax was suggested by @lhoguin. Thank you!