Revamp public API
Now, `khepri` is the main entry point for everything public. There is no
longer a high-level API in `khepri`  and a lower-level API in
`khepri_machine`. `khepri` provides more user-friendly functions and we
will add more when we see fit. Here are a few examples:

* All functions accept both native paths (`[stock, wood, <<"oak">>]`)
  and Unix-like paths (`"/:stock/:wood/oak"`). Note that the syntax for
  Unix-like paths has changed: atoms are prefixed with `:` and binaries
  are left as-is.

* `khepri:get_data(StoreId, PathPattern)` allows to quickly get the data
  attached to a specific node. It's easier than going through
  `khepri:get(StoreId, PathPattern)` and extracting the data from the
  returned map.

Inside transactions, `khepri_tx` provides the same API as `khepri`,
except when the function does not make sense in the context of a
transaction. Unix-like paths are also accepted by `khepri_tx` functions.

`khepri_cluster` is a new module to expose the clustering part of the
API. This was in `khepri` before and was moved to this module. It is
also part of the public interface.

`khepri_path` and `khepri_condition` remain part of the public API for
those needing to manipulate paths.

Other modules are private however. They remain visible in the
documentation because understaing the internals may help sometimes. But
changes in those modules may not be documented in release notes and may
not be reflected in the release versions.