Test: Make sure Coveralls' flag name is unique
This is required by Coverall's API and we didn't follow that so far in
GitHub Actions: only the Erlang version was used in the flag name, but
we run tests against both Linux and Windows for a given version of
Erlang.

Now, the flag name contains both the version of Erlang and the OS
running the testsuite.

Hopefully this will be enough to get rid of the various "0% coverage" we
get lately...