khepri_machine: Improve error report when matching many nodes is denied
The previous error was `{error, matches_many_nodes}`. This was confusing
because it could make the caller think that the command actually found
many nodes, even though the path matched a single one or none at all.

The new error is:

    {error, {possibly_matching_many_nodes_denied, Condition}}

Therefore it highlights that the possibility of matching many is denied,
not the actual result. In addition, it indicates the condition in the
path which caused the problem.