Make `#kpayload_*{}` records private
The public API is now:
* `khepri_payload:none/0`
* `khepri_payload:data/1`
* `khepri_payload:sproc/1`

That said, the caller should rarely have to use them: `khepri:put/1`
tries to automatically detect if and how to wrap the given value.
Therefore, one should be able to just use:

    khepri:put(StoreId, Path, #my_record{...}).

Or:

    khepri:put(StoreId, Path, fun() -> ... end).

Only the "no payload" case has no value beside the internal one. In this
case, the caller should use:

    khepri:put(StoreId, Path, khepri_payload:none()).

Or:

    khepri:clear_payload(StoreId, Path).

The end goal is to reduce the chance that if we change the payload
records, users are forced to recompile their code. It also makes the
code maintenance easier for us.

Note that in the process, the records are renamed `#p_*{}`.