Make `#kevf_*{}` records private
The public API is now provided by the new `khepri_evf` module to
construct the event filters.

Before:

    EventFilter = #kevf_tree{path = Path, props = Props}.

Now:

    EventFilter = khepri_evf:tree(Path, Props).

The reason is the same as for the payload records: if we change the
event filter records, users are not forced to recompile their code.

The `khepri_machine:register_trigger/5` now tries to autodetect and
convert common types to event filter, thanks to `khepri_evf:wrap/1`.
This is the case for native and Unix-like paths which are converted to
tree event filters. This allows users to skip the call to `khepri_evf`
and make it easier to use the API with a path directly.