khepri_tx: allow {atom,safe} variant of OTP-25 call_fun2/3
OTP-25 `call_fun2/3`

    {call_fun2, Tag, Arity, Func}

Calls the function given in the type-tagged register `Func` using
the X-registers `{x,0}` through `{x,Arity-1}` as arguments. Instead
of a function label (for example `{f,123}`), `Func` takes the shape
`{tr,Reg,Type}` (the `tr` record from `beam_asm.hrl`).

In the test case added in this commit, `Func`'s register is populated
by a `move/2`:

    Fun = fun erlang:min/2.
    %% becomes:
    {move,{literal,fun erlang:min/2},{x,0}}

In a child commit we will inspect literal functions in `move/2`s
to ensure an impure function is not passed in. Functions passed in
from the environment are already scanned and extracted.

The `{atom,safe}=Tag` variant of `call_fun2/3` is one of three possible
values:

* `{atom,safe}`: the arity of `Func` is known ahead of time
* `{atom,unsafe}`: nothing about the function is known ahead of time
    * in practice, I have only seen `{atom,unsafe}` within `try`
      expressions
* `{f,Label}`: the function is local

We only cover the `{atom,safe}` variant in this commit.