khepri_fun: unwrap call_fun2/3 Func typing
`call_fun2/3` (see the parent commit) is mistranslated to include
an extra tuple wrapping the typing of `Func`'s register.

    %% from `erlc -S`:
    {call_fun2,{atom,safe},2,{tr,{x,2},{t_fun,2,any,any}}}.
    %% from `beam_disasm`:
    {call_fun2,
     {atom,safe},
     2,
     {tr,{x,2},{{t_fun,any,any,any},0,18446744073709551615}}.

In this commit we discard that wrapping tuple. Determining the arity
of `Func` does not seem necessary: the compiler accepts the `any`
value. On the other hand it would be easy to pull the arity from
the third element of the `call_fun2` tuple.