khepri_utils: Use standard Erlang formatting for node names in displayed tree
The code used to use the same syntax as Unix-like paths. This syntax was
modified, but anyway, it makes more sense in the context of
`khepri:info/1` to present the actual Erlang types, not the Unix-like
path representation. This tool is useful to developers and is not made
for end-users.