khepri_path: Accept Erlang binaries for Unix-like paths
This is especially handy for Beam languages which use Erlang binaries to
implement strings, such as Elixir or Gleam.

To achieve that, `khepri_path:from_string/1` is expanded to accept
Erlang binaries. The implementation is very basic: the binary is
converted to a list.

We also introduce `from_binary/1` which just calls `from_string/1`. So
they both accept all representations of paths. `from_binary/1` is for
convenience and to make it clear that Erlang binaries are accepted.

`to_string/1` will always return an Erlang string however. To produce a
binary, there is now `to_binary/1`.

Fixes #92.