khepri: Redefine the `path()` type
We are now explicit on which types of path are accepted in which places.

The types are:
* `native_path()`, the internal representation made of atoms, binaries
  and, for patterns only, conditions.
* `unix_path()`, the Unix-like path representation, which can be either
  an Erlang string (list of Unicode characters) or an Erlang binary
  (which can be any UTF encoding).

The `path()` type is now the union of both `native_path()` and
`unix_path()`.

Likewise for path patterns with `native_pattern()`, `unix_pattern()` and
`pattern()`. Note that in the case of patterns, only a few conditions
can be represented in the Unix-like path pattern.