khepri_condition: Fix definition of `keep_while` options
Before it was defined as the `keep_whild_conds_map()` internal structure
used to store all `keep_while` conditions in the `khepri_machine` state
machine's state.

The correct type was only `keep_while()`, the one mapping a path and a
condition. The previously defined type didn't match the documentation
and examples.