silence xref warnings from Elixir
Using Mix to compile khepri gives a lot of warnings because the
Khepri mix project calls functions within Elixir (like
`Kernel.defmacro/2`). `:elixir` is not an included app when the
Mix project is configured to be `language: :erlang`, so we have to
explicitly state that we're using Elixir APIs by adding it to the
`:extra_applications` list.