add ExUnit tests for the Elixir API
The `test/test_helper.exs` file is used to start the testing framework.
The most minimal test helper just starts up ExUnit.

`KhepriTest` is effectively empty except for the `doctest Khepri`
line which goes into the `## Examples` headers for any public functions
and tests that the `iex>` line matches the expected output. So

    iex> foo()
    :bar

essentially tests

    assert foo() == :bar

The `import Khepri` line is important: it imports the sigil functions
which allows them to be used in the doctests.