mix.exs: Read application details from Rebar files
The information we need is:
* the application name
* its description
* its version
* its list of dependencies and their version pinning

Everything is available from `src/khepri.app.src` and `rebar.config`.

We do a very basic read of those files. In other words, we don't bother
with Rebar profiles, `rebar.config.script` interpretation and so on. We
expect a very straightforward list of dependencies, but that should be
fine for our needs.

This should solve the problem of keeping `mix.exs` in sync. Indeed, this
was triggered by the fact that Ra's version pinning in `mix.exs` was
out-of-date for some time now.

V2: Use `Keyword` and `to_string/1` instead of `:proplists` and
    `:erlang.list_to_binary/1` to have a more Elixir vibe