khepri_fun: Extract literal funs in `move` instructions
As part of the `call_fun2` instruction introduced in Erlang/OTP 25, it
is apparently possible that the assembly references an anonymous
function directly, instead of an MFA. Therefore, we need to follow and
extract it like any anonymous functions in the environment.

The extraction is performed when we find such a `move` instruction. The
instruction is then updated with a reference to the extracted standalone
function. At the time we prepare the reference, that's ok if the module
is unavailable in the code server.

When the standalone function is executed later, we simply load all
modules stored in the new `literal_funs` list in the `#standalone_fun{}`
record. At runtime, the references to those literal funs will be
resolved successfully.

The `allowed_higher_order_external_call_test` testcase is modified to
reference a function which should be extracted, unlike `erlang:min/2`.