Introduce `khepri:count/{1,2,3}` function to count nodes in the tree
The new function takes a path pattern and returns the number of tree
nodes matching this patttern. The root node is not included in the
result.

It is implemented like `khepri:get/{1,2,3}`. Therefore, it is not
faster. It will consume less memory though, as the result map is not
constructed.