Rework clustering code
The previous clustering code calling Ra was copy-pasted from RabbitMQ
and was not really adapted to Khepri.

The major behavioral change is that `khepri_cluster` only takes care of
the local Ra server, not remote ones anymore. In particular it doesn't
take a list of initial members to create a cluster. In the same spirit,
the `add_member()` function is replaced by a `join()` function: the
local Ra server joins a remote running Ra cluster.

The clustering API should also be way more user-friendly:

1. The `start()` family of functions still takes the name of a running
   Ra system, but a data directory is also accepted: in this case,
   Khepri will configure a Ra system and store files in this given
   directory.

   The default is to start a new Ra system using `khepri#$NODENAME` as
   the data directory in the current working directory.

2. The `start()` family of functions also accepts either the name of a
   Khepri store (i.e. a Ra cluster name as an atom) or a complete Ra
   server configuration (which contains the cluster name).

3. The default Ra system name or data directory, and the default Khepri
   store name can be configured using application environment variables:

    * `default_ra_system` (which can be a Ra system name or a directory)
    * `default_ra_cluster_name`

4. A `stop()` series of functions is introduced. They will stop the
   local Ra server but won't touch the Ra system, even if Khepri started
   it.

5. The `remove_member()` function is gone for now. It will be replaced
   with a function to forcefully evict a member in the future. To remove
   a Ra server from a cluster, you have to use `reset()` locally to the
   node to remove.

The last important change is that Khepri will now take care of waiting
if the local Ra server is not ready to process commands and queries. For
instance if the Ra cluster is still in the process of electing a leader.
All functions take a timeout argument to limit this wait.