Be consistent with the use of "Store ID" and "Ra cluster name"
We now use "Store ID" everywhere. The notion of "cluster" is only
mentionned when we talk about multiple nodes working together.

This allows us to abstract the concept of the Ra cluster below Khepri.
It also makes it clearer that a store ID must be an atom and not any
other types which could be valid as a Ra cluster name.