khepri_cluster: Add a `this_member/1` helper
We frequently used `node_to_member(StoreId, node())`. This warrants the
addition of a dedicated function for this common use case.

While here, move the state machine init parameters in testsuites to a
couple macros. This centralized the parameters map to a header and will
make it easier in the future to modify that map.