Remove LICENSE file, keeping only LICENSE-*
The LICENSE file had an SPDX identifier but licensee, the tool used by
GitHub to detect licenses, doesn't understand it when I tested manually.
GitHub thus shows "Unknown license" for this file.

The SPDX identifier is also present in the README, so we are not loosing
any information.

With only the LICENSE-* files, licensee detects both licenses correctly:

    $ licensee detect
    License:        NOASSERTION
    Matched files:  LICENSE-Apache-2.0, LICENSE-MPL-2.0
    LICENSE-Apache-2.0:
      Content hash:  ab3901051663cb8ee5dea9ebdff406ad136910e3
      Confidence:    100.00%
      Matcher:       Licensee::Matchers::Exact
      License:       Apache-2.0
    LICENSE-MPL-2.0:
      Content hash:  b4db668fa7573bfdcae74eb51eafc961034f0a61
      Confidence:    100.00%
      Matcher:       Licensee::Matchers::Exact
      License:       MPL-2.0