khepri_cluster: Ensure the remembered RaServerConfig contains id
This was not the case when the local Ra server was restarted
successfully. In this case, the initial `RaServerConfig` was stored in a
persistent_term, not the expanded copy where `id` is added to the map.

The code now stores the expanded `RaServerConfig` even after a restart
of the Ra server. There is also an assertion in `remember_store()` to
ensure the key is present.

This bug caused a function clause exception when a Khepri store was
restarted before joining a remote cluster. Indeed, the `join()` function
calls `do_start_server()` which assumes the presence of that `id` key in
the `RaServerConfig` map.

Here is an example with the modified testsuite but without the bugfix:

    {'EXIT',
        {function_clause,
            [{khepri_cluster,complete_ra_server_config,
                 [#{cluster_name => ...}],
                 []},
             {khepri_cluster,do_start_server,2,
                 [{file, "(...)/khepri/src/khepri_cluster.erl"},
                  {line,416}]},
             {khepri_cluster,reset_and_join_locked,6,
                 [{file, "(...)/khepri/src/khepri_cluster.erl"},
                  {line,718}]},
             {khepri_cluster,join,3,
                 [{file, "(...)/khepri/src/khepri_cluster.erl"},
                  {line,637}]},
             {khepri_cluster,join,2,[]}]}}}