khepri_cluster: Log a specific message if store exited before monitor
When we terminate a Khepri store, we first stop the local Ra server then
we monitor it. The idea is to return only when the server exited and
thus is ready to be restarted.

It happens that the process exited before we monitor it, in which case
the stop reason is set to `noproc`. If this happens, we log a specific
message.

The previous default message was confusing:

    Ra server (...) in store "..." exited: noproc