Test: Try to improve capturing of log messages with domain
Let me remind the context: Ra sets a domain on most of the messages it
logs. Common Test's `cth_log_redirect` hook drops them and there is no
configuration parameter to tweak this behavior.

I believe my previous attempt at this was incomplete because it didn't
capture messages from remote Erlang nodes. This patch tries several
things to override checks for the group leader. It looks like it works
now.

While here, the message format is also configured to show the PID of the
process logging a message. This allows to know which node logged it.