Populate the list of skipped modules (in function extraction) during app startup
So far, we were doing it on the first transaction call and the result
was cached. The code is using `application:load/1` to make sure the list
of modules for an application was available. The problem is that it is a
sync call to the application controller with an infinite timeout.

Therefore, if the transaction is executed as part of the start or the
stop of an application, the application controller might be busy and
this `application:load/1` will block indefinitely.

Now, the same code is executed as part of Khepri's application start.
Also, we use the list of already loaded application returned by
`application:loaded_applications/0` to skip the load of already loaded
applications.

Thanks to this, we reduced, perhaps even eliminated the situation where
we call the application controller and cause a dead lock.

The cached list of modules is removed on Khepri's application stop.