Use Ra with aux commands handled before replies
This version of Ra acts on reply and notify effects after other
side-effects have been executed such as aux commands. This will allow
us to use aux commands to send eviction notices to all replicas of
the Khepri cluster whenever a path in the store is updated.

TODO: replace this commit so it uses a newer Ra version from Hex.