Store values in the query-cache when queried
We introduce a khepri_machine:store_in_cache/3 function which stores
return values in the query cache when a query returns successfully.
There are some limitations on what can be cached. Specifically, we
don't currently cache:

* patterns (only exact paths)
* the absence of a node at a path

Currently we cache a `{Pattern, Data}` key-value pair which makes it
easy to add a cache implementation: you just need any fast key-value
store where keys are unique. We could instead cache sub-trees and run
patterns against these subtrees where possible. In order to do this
though we would need the cache to be a trie instead of a key-value
store which is not possible to represent efficiently in ETS.

For the absence of a node at a path, we could cache this by passing
the PathPattern parameter from khepri_machine:get/3 down to the
store_in_cache/3 function. This is left for future experimentation.