Lookup paths in a query cache in khepri_machine:get/3
This cache is not yet filled by queries and there isn't any eviction
yet from the machine.

A cache exists for each server in the cluster so that local queries
may be executed on the local node. Queries of the leader go directly
against the leader node and queries with a `consistent` favor go
are not looked up in the cache (as they may return a new leader).