Cache non-existent paths
We can cache the case where a path being looked up is not present
in the store. This should make `khepri:exists/1,2,3` faster in the
cases where it returns `false`. In addition, we now no longer attempt
to query the cache when a given Path is a pattern. This saves a bit
on `rpc` calls for patterns.