Include query options in query cache key
This covers an edge-case like the one in the
simple_get:get_non_existing_node_test_/0 test case: the result of
a query is based on the input options. If we specify an option
like `expect_specific_node`, the query result will be different.
Instead of using just the path for the cache key, we now include
any options which may change the return value of the query.