khepri: Return errors in exists(), has_data() and has_sproc()
Before this change, all errors were considered the same and `false` was
returned. Now, only the `{error, {node_not_found, _}}` error is
interpreted and `false` is returned. All other errors are returned as
is.

This helps callers who need to know if the given path is incorrect or if
the Ra cluster is degraded.