cluster_SUITE: Add all Erlang nodes to code coverage
Before this change, only the common_test controller node was considered
in the test coverage. But Cover allows to add more node to its report
and that's what this patch does.

We need to add `tools`, the Erlang application providing `cover` to the
Dialyzer extra apps for Dialyzer to succeed now that we call `cover`
explicitly.