khepri_fun: Use locks to avoid concurrent calls to code:get_object_code/1 for the same fun
Parallel calls to handle a function that is not yet in the cache will
simultaneously call code:get_object_code/1. This server call then
becomes very slow, and later all the processes still have to update the
cache - with exactly the same value. Ensuring that only one process
tries to process the function and update the cache yields much faster
parallel execution times, while the other process then also hit the new
cache value and not the processing function path.