Overhaul the Khepri API again
I was still not satisfied with the previous revisit in
commit 3fc0da97e37b08abf4ad0f9a54dd78b28f7c6714. I found the return
values to be too complicated for the most common use cases.

In this commit, the `khepri` and `khepri_tx` modules expose the same
simple API:

* `khepri:get()` returns `{ok, Payload}` directly. If the node has no
  payload, `{ok, undefined}` is returned. It's possible to use
  `khepri:get_or()` to get a default value instead of `undefined`.

* `khepri:get_many()` and `khepri:get_many_or()` are introduced to get
  many nodes in a single call. The result is of the form `{ok, #{Path =>
  Payload}}`.

* Writes (`khepri:put()`, `khepri:delete()`...) always return `ok` in
  case of success.

* There are also `khepri:put_many()` and `khepri:delete_many()` to
  provide the same distinctions as `khepri:get()` and
  `khepri:get_many()`.

The detailed return values are still available from the `khepri_adv` and
`khepri_tx_adv` modules. They expose the same functions as their simple
API counterparts, but return complete values like before this commit.

There are few small changes I want to make that will come in follow-up
commits.