khepri_fun: Allow external funs in options
Previously, the checksum calculation in
`khepri_fun:to_standalone_fun3/2` assumed that function options
were local functions. It's possible to pass external functions as
fully qualified captures, for example:

    fun khepri_tx:ensure_instruction_is_permitted/1

This change fixes a match error that is raised when passing in external
functions as options. When an external function is passed, that
function's md5 is used as the checksum.