Allow multiple ?make_standalone_fun/1 calls
Previously it was impossible to make multiple standalone funs with the
`?make_standalone_fun/1` macro because `__Fun` would be re-bound in
the second invocation of the macro. This change wraps the macro body
in an immediately executed anonymous function. Anonymous functions
have their own scope which doesn't bleed into the outer lexical scope
like with `begin/end`, `case`, or `if`, so the `__Fun` bind is unique.

This change does not rename `__Fun` to `Fun` though because the input
`Expression` could possibly want to use `Fun` binding from its own
scope. Using `__Fun` rather than `Fun` minimizes the chances of an
unexpected conflict.