Return new properties (except payload) after an update
Before this change, functions such as `khepri_adv:put()` would return
the properties and payload of a node as they were before the update.

With this patch, the payload is still the old one, however other
properties such as the payload version correspond to the node after the
update.

This allows to get the payload version of a freshly modified tree node
without having to query it.