Try to make errors and exceptions consistent
With this commit, all errors and exceptions reported by Khepri should
all have the same form:

    {khepri | khepri_ex, Name, Props}

Errors are returned by the Khepri API to notify the failure of an
operation that is to be expected in a correct use of the library. Things
like a missing tree node or a non-matching path pattern. They are
expected to be handled by the caller. Here is an example of an error:

    {error, {khepri, node_not_found, #{...}}} = khepri:get(...)

Exceptions are raised by Khepri using `erlang:error()` to notify an
abnormal use of the library. Things like incorrect arguments or the
failure to prepare a transaction function. They are not expected to be
handled by the caller, they mostly target the developer. Here is an
example of an exception:

    try
        khepri:get([?STAR])
    catch
        error:{khepri_ex, possibly_matching_many_nodes_denied, #{...}} ->
            ...
    end

Some functions will return errors using `erlang:throw()` because their
return value doesn't leave room to return Khepri errors. This is the
case for instance with `khepri:run_sproc(...)` which returns directly
the value of the executed stored procedure otherwise.