khepri_condition: Rename `re_mp()` and fix its use
`re_mp()` was clear what it was only for me apparently :) While renaming
the type to something more human-friendly nad making it more precise, I
discovered that the specs were buggy (because of the lack of precision
before).

In the end, instead of having a type for the compiled regex, the type
now represents the return value of `re:compile/1`.