Transactions: Accept args to pass to a transaction function
It is now possible to write a transaction like this:

    generic_tx(Wood, Increase) ->
        Path = [stock, Wood],
        {ok, Stock} = khepri_tx:get(Path),
	NewStock = erlang:min(?MAX_STOCK, Stock + Increase),
	ok = khepri_tx:put(Path, NewStock).

    increase_wood_stock(Wood, Increase) ->
        khepri:transaction(
	  ?STORE_ID, fun generic_tx/2, [Wood, Increase]).

This way, one can write a more generic transaction function which
doesn't rely on variables in its environment, and pass arguments to
`khepri:transaction()` instead.

This will allow Khepri to use stored procedures as transaction
functions. This will be implemented in a follow-up commit.

References #141.