Acknowledge trigger executions asynchronously
khepri_machine:ack_triggers_execution/2 currently uses a synchronous
command to trigger executions but this is unnecessary: triggers are
executed asynchronously (via gen_server:cast/2) and there aren't any
timing concerns that would require a blocking call.

Synchronous commands may fail and crash the khepri_event_handler
process. In the triggers EUnit suite, calls to
khepri_machine:ack_triggers_execution/2 may race with
test_ra_server_helpers:cleanup/1 and the synchronous command might
cause an exit within khepri_event_handler because the machine is
shutting down. This can cause flaky tests in that suite on hardware
that is fast enough to lose the race.

In this change, we switch to an asynchronous Ra command for
acknowledging trigger executions. Asynchronous Ra commands wil not
fail when the machine is shutting down, so they will not crash the
khepri_event_handler process.