khepri_machine: Automatically reclaim useless tree nodes created automatically
For instance, if a caller stores a value in "/:stock/:wood/oak", both
`stock` and `wood` tree nodes are created on the fly, with no payload
attached to them.

Now, for these automatically created nodes, a "keep while" condition is
added so that they are automatically reclaimed as soon as they have no
child nodes and no payload.

In this example, if "/:stock/:wood/oak" is deleted and no other tree
nodes were added below "/:stock" or "/:stock/:wood", both `stock` and
`wood` will be deleted as well.

Fixes #140.