Add import/export feature
With this patch, it is possible to export and import a Khepri store.

`khepri:export/{2,3,4}` uses a callback module and private data to
export the content of the store. It is possible to specify a path
pattern to filter what should be exported.

`khepri:import/{2,3}` uses a callback module and private data — of
course the same used for export — to import what was previously
exported. This function doesn't remove existing tree nodes. It is up to
the caller to take care of this. There is no support for filtering on
import.

The callback module API is the same as Mnesia's "Backup and restore"
API [1]. The callback module itself doesn't know what is inside the list
of "backup items" passed to it, it's completely opaque.

The implementation uses `khepri_machine` state machine commands as the
export format: each tree node matching the export path pattern is
converted to a `put` command. Tree nodes which don't have any payload or
"keep while" conditions are skipped (i.e. not exported).

During import, the commands are processed one by one. In a future patch,
we could process them in batches to speed things up.

[1] https://www.erlang.org/doc/apps/mnesia/mnesia_app_a

Fixes #153.