Implement ra_machine:overview/1 for khepri_machine
The overview callback formats the state of the machine for functions
like `sys:get_status/1`. The actual state of the machine is still
accessible with `sys:get_state/1`.

The status output was not very readable because of the beam code
binaries in `standalone_fun` records and the use of records for nodes
in the internal representation of the tree. This `overview/1`
implementation uses the map representation of the tree from
`khepri_utils:flat_struct_to_tree/1` and converts compiled standalone
functions into regular functions with `erlang:make_fun/3` (which are
printed with a more concise and readable format).