khepri_fun: Fix `call_fun2` instruction jump label support
Since #149, we resolve and adjust the jump label inside the `call_fun2`
instruction. This way we are as close as possible to the code originally
produced by the compiler and benefit from its optimizations.

However, there is one situation where we can't keep the jump label: it
is when the called local function is not local anymore after extraction.

To avoid many almost identical copies of the same anonymous function
where only the environment changes, `khepri_fun` will extract the
function in a dedicated module and keep the environment separate. If
there is another anonymous function inside this environment, it will
become its own module.

Therefore, it's possible that in the original Beam code, the two
anonymous functions were inside the same module. If the compiler emitted
a `call_fun2` instruction, it could use a jump label in the outer
function to call the inner function passed in the environment.

After extraction by `khepri_fun`, this is no longer possible because the
inner function will be in another module. `call_fun2` must use the
`{atom, safe}` argument instead.