Depend on Ra 2.4.0
Ra 2.4.0 introduces new reply modes which control which member of a
cluster performs the `gen_statem:reply/2` to a command. This can be
used to block the caller until the command has been handled by a given
member. The child commit will introduce a way to pass the new reply
modes in `khepri:command_options()`.