Add 'reply_from' command option
This option wraps the equivalent option in `ra:process_command/3` -
introduced in the dependency upgrade in the parent commit - which
controls which member of the cluster calls `gen_statem:reply/2`. This
can be used to block until a command has been handled by the local
member which can prevent consistency issues with projections. A caller
of `khepri:put/4` might assume that a subsequent `ets:lookup/2` (or
similar) call is up-to-date immediately. By calling `khepri:put/4` with
the `#{reply_from => local}` option, the caller can rely on this
behavior.