Add projections
Projections are similar in spirit to database views. They are a subset
of the store collected into an ETS table which is replicated to all
members of the Khepri cluster. Implementation-wise, projections are
similar to triggers except that projections are updated synchronously
by the machine and are triggered on all changes to a given path
pattern.

Projections can be used for cases where reads are much more common than
writes or in cases where read throughput and/or latency are important.

This change initially introduces projections with the following
additions to the API:

* `khepri_projection:new/3` - Create a projection record.
* `khepri_projection:name/1` - Retrieve the name (the ETS table name)
  of the projection. (Necessary because the projection record is
  opaque).
* `khepri:register_projection/4` - Register a projection within the
  store against a given path pattern.

As well as the internal implementation in `khepri_machine` and the
associated documentation.